+++ src/constraint_handler/main.py
@@ -11,6 +11,7 @@ import constraint_handler.propagator as propagator
 import constraint_handler.schemas.atom as atom
 import constraint_handler.solver_environment as solver_environment
 import flat_ch.main as flat_main
+from constraint_handler.engine import Engine, compile
 
 module_main = [
     "main",
@@ -156,6 +157,7 @@ def add_to_control(
     environment=None,
     _environment_ids=dict(),
     api: str = "ch",
+    engine: Engine = compile,
 ):
     """Adds encoding logic to the provided Control instance. The environment argumennt specifies the locals used in the python statements and expressions."""
     if api == "fch":
@@ -187,6 +189,7 @@ def add_to_control(
             evaluator._solver_environment[idx] = environment
             _environment_ids[eid] = idx
         ctrl.add(f"main_solverIdentifier({idx}).")
+    ctrl.add(engine.program())
     setup_propagator(ctrl)